replace enum with class for each menu
