+++ src/main/java/de/hysky/skyblocker/mixins/HandledScreenProviderMixin.java
@@ -10,6 +10,8 @@ import de.hysky.skyblocker.skyblock.dungeon.partyfinder.PartyFinderScreen;
 import de.hysky.skyblocker.skyblock.item.SkyblockCraftingTableScreenHandler;
 import de.hysky.skyblocker.skyblock.item.SkyblockCraftingTableScreen;
 import de.hysky.skyblocker.skyblock.radialMenu.RadialMenu;
+import de.hysky.skyblocker.skyblock.radialMenu.RadialMenuManager;
+import de.hysky.skyblocker.skyblock.radialMenu.RadialMenuScreen;
 import de.hysky.skyblocker.skyblock.tabhud.config.WidgetsConfigurationScreen;
 import de.hysky.skyblocker.utils.Utils;
 import net.minecraft.client.MinecraftClient;
@@ -24,7 +26,6 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import java.util.Arrays;
 
 @Mixin(HandledScreens.Provider.class)
 public interface HandledScreenProviderMixin<T extends ScreenHandler> {
@@ -118,11 +119,11 @@ public interface HandledScreenProviderMixin<T extends ScreenHandler> {
 			}
 
 			// radial menus
-			case GenericContainerScreenHandler containerScreenHandler when Arrays.stream(RadialMenu.MenuType.values()).anyMatch(menuType -> menuType.match(nameLowercase)) -> {
+			case GenericContainerScreenHandler containerScreenHandler when RadialMenuManager.doseMenuExistForTitle(nameLowercase) -> {
 				client.player.currentScreenHandler = containerScreenHandler;
-				RadialMenu.MenuType menuType = Arrays.stream(RadialMenu.MenuType.values()).filter(t -> t.match(nameLowercase)).findFirst().get();
+				RadialMenu menuType = RadialMenuManager.getMenuForTitle(nameLowercase);
 
-				client.setScreen(new RadialMenu(containerScreenHandler, menuType, name));
+				client.setScreen(new RadialMenuScreen(containerScreenHandler, menuType, name));
 
 				ci.cancel();
 			}
