+++ src/main/java/de/hysky/skyblocker/skyblock/radialMenu/RadialMenu.java
@@ -1,159 +1,30 @@
 package de.hysky.skyblocker.skyblock.radialMenu;
 
-import it.unimi.dsi.fastutil.ints.Int2IntMap;
-import it.unimi.dsi.fastutil.ints.Int2IntOpenHashMap;
-import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
-import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.gui.DrawContext;
-import net.minecraft.client.gui.screen.Screen;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-import net.minecraft.screen.GenericContainerScreenHandler;
-import net.minecraft.screen.ScreenHandler;
-import net.minecraft.screen.ScreenHandlerListener;
-import net.minecraft.screen.slot.SlotActionType;
 import net.minecraft.text.Text;
 
-import javax.annotation.Nullable;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.regex.Pattern;
+public interface RadialMenu {
+	/**
+	 * @param title the original window title
+	 * @return  the title show in the radial menu
+	 */
+	Text getTitle(Text title);
+
+	/**
+	 * Works out if this radial menu should be used for this screen title
+	 * @param title screen title to compare
+	 * @return the menu should be used
+	 */
+	boolean titleMatches(String title);
+
+	/**
+	 * Works out if an item should be added to the menu
+	 * @param slotId the slot the items in
+	 * @param stack the item
+	 * @return if it should be added
+	 */
+	boolean itemMatches(int slotId, ItemStack stack);
 
 
-public class RadialMenu extends Screen implements ScreenHandlerListener {
-	private static final MinecraftClient CLIENT = MinecraftClient.getInstance();
-	private final MenuType menuType;
-	private final Int2ObjectOpenHashMap<ItemStack> options = new Int2ObjectOpenHashMap<>();
-	private final Int2IntMap syncIds = new Int2IntOpenHashMap();
-	private final List<RadialButton> buttons = new ArrayList<>();
-	private final GenericContainerScreenHandler handler;
-
-
-	public RadialMenu(GenericContainerScreenHandler handler, MenuType type, Text title) {
-		super(type.showTitle == null ? title : Text.of(type.showTitle));
-		menuType = type;
-
-		options.clear();
-		//Listen for slot updates
-		this.handler = handler;
-		handler.addListener(this);
-	}
-
-	@Override
-	protected void init() {
-		super.init();
-
-		//create buttons
-		buttons.clear();
-		clearChildren();
-		float angle = 0;
-		float buttonArcSize = (float) ((2 * Math.PI) / options.size());
-		List<Int2ObjectMap.Entry<ItemStack>> optionOrdered = new ArrayList<>(options.int2ObjectEntrySet().stream().toList());
-
-		//check for back and close buttons to put at bottom
-		Int2ObjectMap.Entry<ItemStack> backSlot = optionOrdered.stream().filter(option -> validName(option.getValue(), "Go Back")).findAny().orElse(null);
-		Int2ObjectMap.Entry<ItemStack> closeSlot = optionOrdered.stream().filter(option -> validName(option.getValue(), "Close")).findAny().orElse(null);
-		int bottom = (int) Math.ceil((optionOrdered.size() - ((closeSlot == null) ? 0 : 1) - ((backSlot == null) ? 0 : 1)) / 2f);
-
-		optionOrdered.remove(backSlot);
-		optionOrdered.remove(closeSlot);
-		if (backSlot != null) {
-			optionOrdered.add(bottom, backSlot);
-		}
-		if (closeSlot != null) {
-			optionOrdered.add(bottom, closeSlot);
-		}
-
-
-		for (Int2ObjectMap.Entry<ItemStack> stack : optionOrdered) {
-			RadialButton newButton = new RadialButton(angle, buttonArcSize, 50, 100, stack.getValue(), button -> this.clickSlot(stack.getIntKey(),button), stack.getIntKey());
-			buttons.add(newButton);
-			addDrawableChild(newButton);
-			angle += buttonArcSize;
-		}
-
-	}
-
-	private static boolean validName(ItemStack stack, String validName) {
-		Text customName = stack.getCustomName();
-		if (customName == null) return false;
-		return customName.getString().equals(validName);
-
-	}
-
-	private void clickSlot(int slotId, int button) {
-		if (CLIENT.interactionManager == null || !syncIds.containsKey(slotId)) return;
-		CLIENT.interactionManager.clickSlot(syncIds.get(slotId), slotId, button, SlotActionType.PICKUP, CLIENT.player);
-	}
-
-	@Override
-	public void onSlotUpdate(ScreenHandler handler, int slotId, ItemStack stack) {
-
-		if (this.handler == null || slotId >= this.handler.getRows() * 9) return;
-
-
-		if (menuType.itemMatches(slotId, stack)) {
-			options.put(slotId, stack);
-			syncIds.put(slotId, handler.syncId);
-			init();
-		}
-	}
-
-	@Override
-	public void render(DrawContext context, int mouseX, int mouseY, float deltaTicks) {
-		super.render(context, mouseX, mouseY, deltaTicks);
-
-		//render menu title
-		context.drawCenteredTextWithShadow(textRenderer, getTitle(), width / 2, height / 2 - textRenderer.fontHeight, 0xFFFFFF);
-		//draw separation line
-		int textWidth = textRenderer.getWidth(getTitle());
-		context.drawHorizontalLine(width / 2 - textWidth / 2, width / 2 + textWidth / 2, height / 2, 0xFFFFFFFF);
-		//render current option name
-		buttons.stream().filter(button -> button.hovered).findAny().ifPresent(hovered ->
-				context.drawCenteredTextWithShadow(textRenderer, hovered.getName(), width / 2, height / 2 + 2, 0xFFFFFFFF)); // + 2 to move out of way of line
-	}
-
-
-	@Override
-	public void onPropertyUpdate(ScreenHandler handler, int property, int value) {
-
-	}
-
-	public enum MenuType {
-		YOURBAGS("your bags", null, null, List.of(Items.BLACK_STAINED_GLASS_PANE)),
-		YOURSKILLS("your skills", null, null, List.of(Items.BLACK_STAINED_GLASS_PANE)),
-		FASTTRAVEL("(fast travel)|(.* warps)", null, null, List.of(Items.BLACK_STAINED_GLASS_PANE)),
-		HUNTINGTOOLKIT("hunting toolkit ➜ selection", "Hunting Toolkit", List.of(41), List.of(Items.BLACK_STAINED_GLASS_PANE,Items.GRAY_DYE)),
-		SKYBLOCKMENU("skyblock menu", null, null, List.of(Items.BLACK_STAINED_GLASS_PANE));
-
-		final Pattern title;
-		final String showTitle;
-
-		MenuType(String title, @Nullable String showTitle, @Nullable List<Integer> blackListSlots, @Nullable List<Item> blackListItems) {
-			this.title = Pattern.compile(title);
-			this.showTitle = showTitle;
-			this.blackListSlots = blackListSlots;
-			this.blackListItems = blackListItems;
-		}
-		final List<Integer> blackListSlots;
-		final List<Item> blackListItems;
-
-		public boolean match(String name) {
-			//return false;
-			return this.title.matcher(name).matches();
-		}
-
-		public boolean itemMatches(int slotId, ItemStack stack) {
-			if (blackListItems != null) {
-				if (blackListItems.contains(stack.getItem())) return false;
-			}
-			if (blackListSlots != null) {
-				if (blackListSlots.contains(slotId)) return false;
-			}
-
-			return true;
-		}
-	}
+	boolean isEnabled();
 }
