+++ src/main/java/de/hysky/skyblocker/skyblock/radialMenu/RadialMenuManager.java
@@ -1,0 +1,23 @@
+package de.hysky.skyblocker.skyblock.radialMenu;
+
+import de.hysky.skyblocker.skyblock.radialMenu.menus.FastTravelMenu;
+import de.hysky.skyblocker.skyblock.radialMenu.menus.HuntingMenu;
+import de.hysky.skyblocker.skyblock.radialMenu.menus.SkyblockMenu;
+
+import java.util.Arrays;
+
+public class RadialMenuManager {
+	private static final RadialMenu[] menus = new RadialMenu[]{
+			new FastTravelMenu(),
+			new HuntingMenu(),
+			new SkyblockMenu()
+	};
+
+	public static boolean doseMenuExistForTitle(String title){
+		return Arrays.stream(menus).anyMatch(menu-> menu.titleMatches(title));
+	}
+
+	public static RadialMenu getMenuForTitle(String title){
+		return Arrays.stream(menus).filter( menu -> menu.titleMatches(title)).findFirst().orElse(null);
+	}
+}
