+++ src/main/java/de/hysky/skyblocker/skyblock/radialMenu/menus/SkyblockMenu.java
@@ -1,0 +1,7 @@
+package de.hysky.skyblocker.skyblock.radialMenu.menus;
+
+public class SkyblockMenu extends BasicMenu {
+	public SkyblockMenu() {
+		super("skyblock menu");
+	}
+}
